Implement a Bolt-backed job store
Note that this hasn't actually been tested at all.

+++ job.go
@@ -24,8 +24,8 @@ type JobRequest struct {
 
 // Job represents the job data that should be persisted to a JobStore.
 type Job struct {
-	ID     uint
-	Number uint
+	ID     uint64
+	Number uint64
 
 	Owner  string
 	Repo   string
